requirements.txt
# Telegram Bot
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Bitrix24
requests==2.31.0

# Google Drive
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Scheduler
apscheduler==3.10.4

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
